Module Ctypes_structs_computed

Structs and unions whose layouts are computed from the sizes and alignment requirements of the constituent field types.

include Ctypes_structs.S with type ('a, 's) field := ('a, 's) Ctypes_static.field

val field : 't Ctypes_static.typ -> string -> 'a Ctypes_static.typ -> ('a, ('s, [< `Struct | `Union ]) Ctypes_static.structured as 't) Ctypes_static.field

val seal : (_, [< `Struct | `Union ]) Ctypes_static.structured Ctypes_static.typ -> unit